requirements.txt
pip-chill==1.0.3
python-dotenv==1.0.1
telebot==0.0.5
torchvision==0.17.2
transformers==4.39.3
